requirements.txt
datasets==4.5.0
matplotlib==3.9.0
numpy==2.4.2
opencv-python==4.8.1.78
Pillow==12.1.1
scikit_learn==1.8.0
timm==1.0.24
torch==2.8.0
torchvision==0.23.0
tqdm==4.67.3
transformers==4.57.6
